+++ src/main/java/triple/Vertex.java
@@ -5,13 +5,13 @@ import index.MySerialzable;
 public class Vertex implements MySerialzable {
 
 
-    public long v;
-    public  long e;
+    public int v;
+    public  int e;
     public int weight1;
     public int weight2;
     public boolean isBorder;
 
-    public Vertex(long v, long e) {
+    public Vertex(int v, int e) {
         this.v = v;
         this.e = e;
         weight1 = 1;
@@ -39,8 +39,8 @@ public class Vertex implements MySerialzable {
     public Vertex deSerialize(String res){
         Vertex vertex = new Vertex(0,0);
         String [] resArr = res.split(",");
-        vertex.v = Long.valueOf(resArr[0]);
-        vertex.e = Long.valueOf(resArr[1]);
+        vertex.v = Integer.valueOf(resArr[0]);
+        vertex.e = Integer.valueOf(resArr[1]);
         vertex.weight1 = Integer.valueOf(resArr[2]);
         vertex.weight2 = Integer.valueOf(resArr[3]);
         vertex.isBorder= Integer.valueOf(resArr[4]) == 1;
